Manage UIDs in the system on creation
Also emit a warning message when a uid is changed for an object.
This occurs, for instance, in the stale reads case when the node writes
the pod update for the old pod but uses the old uid, not the new one.
This could be rejected but currently is not.

For the causal model we might need to do something different for this to
handle conflicts, maybe changing the resources to be indexed by uid
rather than name.