Handle resource_version setting in state
This would normally be done on retrieving resources from etcd.
However, in our case we don't have the round trip so we should just
store the mod revision when we insert it (we know it already).

This happens on all resources and can block insertion when the resource
version is specified but doesn't match (would be a 409 conflict in k8s).

To handle this we also make the apply_changes function more
transaction-like by having an inner version that can return errors and
we only use the state and update the revision if it is successful.