Split and update job status count properties
Also remove property for active pods having finalizers.
This is a tricky property as an active pod with a finalizer can't be
deleted in k8s, so the finalizer needs to be removed. Deleting it first
could expect the job controller to observe it and remove the finalizer
but that isn't how it operates, ignoring non-active pods.